Add files via upload
This GitHub repository hosts a comprehensive Breast Cancer Prediction project utilizing machine learning techniques. The project involves preprocessing a diverse dataset of clinical features and genetic markers. A variety of state-of-the-art algorithms, such as Random Forest, Support Vector Machine, and Neural Networks, are implemented and compared for predictive accuracy. The repository provides well-documented code, enabling researchers and practitioners to understand, reproduce, and build upon the study. The project not only showcases the application of machine learning in healthcare but also contributes to early detection and personalized treatment strategies for breast cancer. Through transparent methodologies and insightful results, this repository serves as a valuable resource for those interested in cancer prediction, fostering collaboration and innovation in the field of medical data analysis.